app/requirements.txt
flask==2.0.3
werkzeug==2.0.3
requests==2.28.2
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-exporter-otlp==1.12.0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-instrumentation-requests==0.33b0
prometheus-client==0.16.0
